requirements.txt
anyio==3.7.1
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
Django==4.2.3
django-crispy-forms==2.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
wikipedia==1.4.0
youtube-search-python==1.6.6
